Some refactoring to clean things up, get rid of some dead code
